% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docker_running.R
\name{check_docker_running}
\alias{check_docker_running}
\title{check_docker_running}
\usage{
check_docker_running()
}
\arguments{
\item{POSTGRES_HOST}{the host}
}
\description{
Function checks if docker is running and returns a message
with information on any currently active containers
}
