% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{connect_to_db}
\usage{
connect_to_db(
  POSTGRES_HOST = NULL,
  POSTGRES_PORT = NULL,
  POSTGRES_USER = NULL,
  POSTGRES_PASSWORD = NULL,
  POSTGRES_DB = NULL,
  n_connection_tries = 5
)
}
\arguments{
\item{POSTGRES_HOST}{the host}

\item{POSTGRES_PORT}{the port number}

\item{POSTGRES_USER}{the user}

\item{POSTGRES_PASSWORD}{the password}

\item{POSTGRES_DB}{the database name}

\item{n_connection_tries}{number of attempts to connect to the database. default is 5}
}
\description{
Connect to the postgres database that lives in docker.
}
\details{
The default value for each parameter is taken from the .Environ file.
}
