% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_data.R
\name{get_available_years}
\alias{get_available_years}
\title{Get the valid years of data for a given player_id}
\usage{
get_available_years(player_id)
}
\arguments{
\item{player_id}{a valid player id value. Use \code{get_current_owgr} to
get a list of valid player_id values}
}
\description{
Get the valid years of data for a given player_id
}
