% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_owgr.R
\name{get_current_owgr}
\alias{get_current_owgr}
\title{Function to extract the current official world golf rankings}
\usage{
get_current_owgr()
}
\description{
Function to extract the current official world golf rankings
}
\details{
This function extracts the current Official World
Golf Rankings from www.owgr.com. A tibble is returned which
contains the current rankings along with links to player
specific pages. These links can be used as input to
player data extraction functions.
}
\examples{
get_current_owgr()

}
